pyproject.toml
[project]
name = "a2a-text-utilities"
version = "0.1.0"
description = "A2A Text Utilities Agent — demo agent for Azure deployment"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.2.9",
    "azure-identity>=1.19.0",
    "azure-monitor-opentelemetry>=1.6.0",
    "httpx>=0.28.1",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "client"]
